+++ src/app/modules/podcast/podcast.component.ts
@@ -1,10 +1,13 @@
 import { Component, OnInit } from '@angular/core';
-import { ActivatedRoute } from '@angular/router';
+import { ActivatedRoute, Router } from '@angular/router';
 import { Store } from '@ngrx/store';
 import { Observable, distinctUntilChanged, filter, map } from 'rxjs';
-import { PodCastDetail } from 'src/app/shared/models/podcast-detail';
+import { PodCastDetail, Track } from 'src/app/shared/models/podcast-detail';
 import { findPodCast } from 'src/app/store/podcast.actions';
-import { selectCurrentPodCast } from 'src/app/store/podcast.selector';
+import {
+  selectCurrentPodCast,
+  selectIsLoading,
+} from 'src/app/store/podcast.selector';
 
 @Component({
   selector: 'app-podcast',
@@ -13,10 +16,12 @@ import { selectCurrentPodCast } from 'src/app/store/podcast.selector';
 })
 export class PodCastComponent implements OnInit {
   podCast$: Observable<PodCastDetail> = this.store.select(selectCurrentPodCast);
+  _id: string;
 
   constructor(
     private store: Store<{}>,
-    private activatedRoute: ActivatedRoute
+    private activatedRoute: ActivatedRoute,
+    private router: Router
   ) {}
 
   ngOnInit(): void {
@@ -27,7 +32,12 @@ export class PodCastComponent implements OnInit {
         map((params) => params['id'])
       )
       .subscribe((id: string) => {
+        this._id = id;
         this.store.dispatch(findPodCast({ data: id }));
       });
+  }
+
+  redirectToEpidose(track: Track): void {
+    this.router.navigate([`podcast/${this._id}/episode/${track.trackId}`]);
   }
 }