+++ src/app/store/podcast.actions.ts
@@ -1,6 +1,7 @@
 import { createAction, props } from '@ngrx/store';
 import { PodCastFeed } from '../shared/models/podcast';
-import { PodCastDetail } from '../shared/models/podcast-detail';
+import { PodCastDetail, Track } from '../shared/models/podcast-detail';
+import { TransmissionState } from './podcast-store.model';
 
 export const init = createAction('[PodCast] -  INIT', props<{ data: any }>());
 
@@ -18,4 +19,19 @@ export const findPodCast = createAction(
 export const findPodCastSuccess = createAction(
   '[PodCast] -  FIND PODCAST SUCCESS',
   props<{ data: PodCastDetail }>()
+);
+
+export const findPodCastEpisode = createAction(
+  '[PodCast] -  FIND PODCAST EPISODE',
+  props<{ data: string }>()
+);
+
+export const findPodCastEpisodeSuccess = createAction(
+  '[PodCast] -  FIND PODCAST EPISODE SUCCESS',
+  props<{ data: PodCastDetail; track: Track }>()
+);
+
+export const setTransmission = createAction(
+  '[PodCast] -  SET TRANSMISSION',
+  props<{ data: TransmissionState }>()
 );